refactor(wallet): move useCollectible hook, dedupe nft-helpers, derive chainId from network
