+++ apps/wallet/src/routes/portfolio/collectibles/-components/nft-helpers.ts
@@ -1,7 +1,10 @@
+import {
+  encodeNftTransfer,
+  extractTxHash,
+  isSupportedNftStandard,
+} from '@status-im/wallet/components'
+import { NETWORK_TO_CHAIN_ID } from '@status-im/wallet/data'
 import { Interface } from 'ethers'
-import { isAddress } from 'viem'
-import { formatEther } from 'viem/utils'
-import * as z from 'zod'
 
 import {
   fetchTrpcData,
@@ -10,18 +13,14 @@ import {
 
 import type { NetworkType } from '@status-im/wallet/data'
 
-const erc721 = new Interface([
-  'function safeTransferFrom(address from, address to, uint256 tokenId)',
-])
+export { extractTxHash, isSupportedNftStandard }
+
+const JSONRPC_REQUEST_ID = 1
+const JSONRPC_VERSION = '2.0'
 
 const erc1155 = new Interface([
-  'function safeTransferFrom(address from, address to, uint256 id, uint256 amount, bytes data)',
   'function balanceOf(address account, uint256 id) view returns (uint256)',
 ])
-
-const NETWORK_TO_CHAIN_ID: Record<string, number> = {
-  ethereum: 1,
-}
 
 const parseTokenId = (tokenId: string): bigint => {
   try {
@@ -29,232 +28,6 @@ const parseTokenId = (tokenId: string): bigint => {
   } catch {
     throw new Error(`Invalid token id: ${tokenId}`)
   }
-}
-
-const ZERO_ADDRESS = '0x0000000000000000000000000000000000000000'
-
-const tryBigInt = (val: string): bigint | null => {
-  try {
-    return BigInt(val)
-  } catch {
-    return null
-  }
-}
-
-export const createNftSendSchema = (params: {
-  fromAddress: string
-  isErc1155: boolean
-  balance?: bigint
-}) => {
-  const { fromAddress, isErc1155, balance } = params
-
-  return z.object({
-    to: z
-      .string()
-      .min(1, 'Recipient address is required')
-      .refine(
-        val => isAddress(val, { strict: true }),
-        'The address is not valid',
-      )
-      .refine(
-        val => val.toLowerCase() !== ZERO_ADDRESS,
-        "Can't send to the zero address",
-      )
-      .refine(
-        val => val.toLowerCase() !== fromAddress.toLowerCase(),
-        "Can't send to the sender address",
-      ),
-    amount: z
-      .string()
-      .refine(
-        val => !isErc1155 || !val.includes('.'),
-        "Can't send fraction of collectible",
-      )
-      .refine(
-        val => !isErc1155 || /^\d+$/.test(val),
-        'Amount must be a positive integer',
-      )
-      .refine(val => {
-        if (!isErc1155) return true
-        const big = tryBigInt(val)
-        return big !== null && big > 0n
-      }, 'Amount must be greater than 0')
-      .refine(val => {
-        if (!isErc1155 || balance === undefined) return true
-        const big = tryBigInt(val)
-        return big !== null && big <= balance
-      }, 'More than available balance'),
-  })
-}
-
-export type NftSendFormData = z.infer<ReturnType<typeof createNftSendSchema>>
-
-export const isSupportedNftStandard = (standard: string) =>
-  ['ERC721', 'ERC1155'].includes(standard)
-
-export const extractTxHash = (id: unknown): string | undefined => {
-  if (typeof id === 'string') {
-    return id
-  }
-
-  if (id && typeof id === 'object') {
-    const obj = id as Record<string, unknown>
-
-    if ('result' in obj && typeof obj.result === 'string') {
-      return obj.result
-    }
-
-    if ('txid' in obj && typeof obj.txid === 'string') {
-      return obj.txid
-    }
-  }
-
-  return undefined
-}
-
-const getErrorMessage = (error: unknown): string | null => {
-  if (error instanceof Error) {
-    return error.message
-  }
-
-  if (typeof error === 'string') {
-    return error
-  }
-
-  if (typeof error === 'object' && error !== null) {
-    if ('message' in error && typeof error.message === 'string') {
-      return error.message
-    }
-
-    if ('error' in error && typeof error.error === 'string') {
-      return error.error
-    }
-  }
-
-  return null
-}
-
-export type NftSendErrorCode =
-  | 'insufficient_funds'
-  | 'user_rejected'
-  | 'not_approved'
-  | 'erc1155_insufficient_balance'
-  | 'execution_reverted'
-  | 'tx_hash_missing'
-  | 'wallet_locked'
-  | 'unknown'
-
-export type NftSendError = {
-  code: NftSendErrorCode
-  message: string
-  data?: { haveWei: bigint; wantWei: bigint }
-}
-
-const insufficientFundsPatterns = [
-  /insufficient funds for gas \* price \+ value: have (\d+) want (\d+)/i,
-  /insufficient funds.*have[:\s]+(\d+).*want[:\s]+(\d+)/i,
-]
-
-export const classifyNftSendError = (error: unknown): NftSendError => {
-  const message = getErrorMessage(error)
-
-  if (!message) {
-    return { code: 'unknown', message: 'Failed to send NFT' }
-  }
-
-  for (const pattern of insufficientFundsPatterns) {
-    const match = message.match(pattern)
-    if (match) {
-      const haveWei = BigInt(match[1])
-      const wantWei = BigInt(match[2])
-      return {
-        code: 'insufficient_funds',
-        message: `Not enough ETH for gas. Have ${formatEther(haveWei)} ETH, need up to ${formatEther(wantWei)} ETH. Short ${formatEther(wantWei - haveWei)} ETH.`,
-        data: { haveWei, wantWei },
-      }
-    }
-  }
-
-  if (/insufficient funds/i.test(message)) {
-    return {
-      code: 'insufficient_funds',
-      message: 'Not enough ETH to pay for gas.',
-    }
-  }
-
-  if (/user rejected|user denied|rejected request|cancelled/i.test(message)) {
-    return { code: 'user_rejected', message: 'Transaction was cancelled.' }
-  }
-
-  if (/not owner nor approved|caller is not token owner/i.test(message)) {
-    return {
-      code: 'not_approved',
-      message: 'This wallet is not allowed to transfer this NFT.',
-    }
-  }
-
-  if (/insufficient balance for transfer/i.test(message)) {
-    return {
-      code: 'erc1155_insufficient_balance',
-      message: 'Not enough balance to transfer this ERC1155 NFT.',
-    }
-  }
-
-  if (/execution reverted/i.test(message)) {
-    return {
-      code: 'execution_reverted',
-      message: 'The NFT contract rejected this transfer.',
-    }
-  }
-
-  if (/transaction hash not found/i.test(message)) {
-    return {
-      code: 'tx_hash_missing',
-      message: 'The node did not return a transaction hash.',
-    }
-  }
-
-  if (/wallet not unlocked/i.test(message)) {
-    return {
-      code: 'wallet_locked',
-      message: 'Wallet is locked. Unlock it and try again.',
-    }
-  }
-
-  return { code: 'unknown', message: 'Failed to send NFT' }
-}
-
-export const getNftSendErrorMessage = (error: unknown): string =>
-  classifyNftSendError(error).message
-
-export function encodeNftTransfer(params: {
-  standard: string
-  from: string
-  to: string
-  tokenId: string
-  amount?: string
-}): string {
-  const id = parseTokenId(params.tokenId)
-
-  if (params.standard === 'ERC1155') {
-    return erc1155.encodeFunctionData('safeTransferFrom', [
-      params.from,
-      params.to,
-      id,
-      BigInt(params.amount ?? '1'),
-      '0x',
-    ])
-  }
-
-  if (params.standard === 'ERC721') {
-    return erc721.encodeFunctionData('safeTransferFrom', [
-      params.from,
-      params.to,
-      id,
-    ])
-  }
-
-  throw new Error(`Unsupported NFT standard: ${params.standard}`)
 }
 
 function buildNftGasFeeParams(params: {
@@ -303,8 +76,8 @@ export async function fetchErc1155Balance(params: {
     method: 'POST',
     headers: { 'Content-Type': 'application/json' },
     body: JSON.stringify({
-      jsonrpc: '2.0',
-      id: 1,
+      jsonrpc: JSONRPC_VERSION,
+      id: JSONRPC_REQUEST_ID,
       method: 'eth_call',
       params: [{ to: params.contract, data }, 'latest'],
     }),
