+++ packages/wallet/src/components/send-collectibles-modal/index.tsx
@@ -340,3 +340,8 @@ const SendCollectiblesModal = (props: Props) => {
 
 export { SendCollectiblesModal }
 export type { GasFees as CollectibleGasFees }
+export {
+  encodeNftTransfer,
+  extractTxHash,
+  isSupportedNftStandard,
+} from './nft-helpers'
